13.5.1 works*, though there's an exiled bug
the unrelated bug in question is exiled erroring before setting the inventory when spawning. im not sure why or how SLCryptoAuth is triggering this exiled bug but it's occurring anyway so it looks like i'll have to separately fork exiled and bring bug fixes from future versions to old versions. the fix is literally just this PR, but it was made way after exiled 8.12.2 https://github.com/ExMod-Team/EXILED/pull/282